fix(tests): update project references to libraries
The project file `Hexalith.Commons.Tests.csproj` has been modified to change the paths of several project references. The previous references to projects in the `src/packages` directory have been removed, and new references to projects in the `src/libraries` directory have been added. This restructuring improves the organization of the project.